requirements.txt
Django==1.10.5
django-ckeditor==5.2.1
olefile==0.44
Pillow==4.0.0
psycopg2==2.6.2
sorl-thumbnail==12.3
